+++ frontend/components/Header/Header.scss
@@ -1,3 +1,3 @@
 .header {
-  background-color: gray;
+  background-color: $oc-gray-3;
 }
