+++ frontend/next.config.js
@@ -1,6 +1,9 @@
 const withTypescript = require("@zeit/next-typescript");
 const withSass = require('@zeit/next-sass');
 
+const isEnvProduction = process.env === "production";
+const shouldUseSourceMap = process.env.GENERATE_SOURCEMAP !== 'false';
+
 module.exports = withTypescript(
   withSass({
     cssModules: true,
@@ -8,11 +11,11 @@ module.exports = withTypescript(
       importLoaders: 1, // 0 => no loaders (default); 1 => postcss-loader; 2 => postcss-loader, sass-loader
       localIdentName: "[local]___[hash:base64:5]",
     },
-    // sassLoaderOptions: {
-    //   includePaths: ['styles'],
-    //   sourceMap: isEnvProduction && shouldUseSourceMap,
-    //   data: `@import 'utils';`
-    // },
+    sassLoaderOptions: {
+      includePaths: [__dirname +'/styles'],
+      sourceMap: isEnvProduction && shouldUseSourceMap,
+      data: `@import 'utils';`
+    },
     webpack(config, options) {
       // Further custom configuration here
       return config;
