+++ frontend/package.json
@@ -27,9 +27,12 @@
     "@types/next": "^8.0.5",
     "@zeit/next-sass": "^1.0.1",
     "@zeit/next-typescript": "^1.1.1",
-    "prettier": "^1.17.0",
     "classnames": "^2.2.6",
+    "include-media": "^1.4.9",
     "node-sass": "^4.12.0",
+    "normalize.css": "^8.0.1",
+    "open-color": "^1.6.3",
+    "prettier": "^1.17.0",
     "sass-lint": "^1.13.1",
     "sass-lint-auto-fix": "^0.17.0",
     "tslint": "^5.16.0",
