+++ frontend/styles/base.scss
@@ -1,0 +1,20 @@
+// 전역으로 사용하는 스타일 폰트 설정
+@import '~normalize.css';
+@import url('//fonts.googleapis.com/earlyaccess/notosanskr.css');
+
+body {
+  -moz-osx-font-smoothing: grayscale;
+  -webkit-font-smoothing: antialiased;
+  margin: 0;
+  box-sizing: border-box;
+  font-family: 'Note Sans KR', sans-serif;
+}
+
+* {
+  box-sizing: inherit;
+}
+
+a {
+  color: inherit;
+  text-decoration: inherit;
+}
