+++ frontend/styles/utils.scss
@@ -1,0 +1,4 @@
+/* 컴포넌트를 스타일링 할 때 사용하는 스타일 유틸리티*/
+@import 'utils/lib';
+@import 'utils/mixins';
+@import 'utils/variables';
