+++ frontend/styles/utils/_variables.scss
@@ -1,0 +1,1 @@
+$breakpoints: (small: 320px, medium: 768px, large: 1024px, wide: 1400px);
